Defined bases folder & deport Root Nav in dedicated file.
